\name{GeneralResults}
\alias{GeneralResults}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Simple widget to dispplay text
}
\description{
Simple widget to dispplay text
}
\usage{
GeneralResults(title = "Results of Analysis", text = "Results", Toolkit = "RGtk2", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{title}{
Title of the window
}
  \item{text}{
Text to display
}
  \item{Toolkit}{
RGtk2. Not tried with other toolkits
}
  \item{\dots}{
Aditional arguments
}
}
\details{
Simple widget to dispplay text
}
\value{
No value returned
}

\author{
Jose Luis Vicente Villardon
}


\examples{
GeneralResults(text="Show me")
}

