\name{MultBiplotRGUI}
\alias{MultBiplotRGUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Graphical user interface for the initial window
}
\description{
Graphical user interface for the initial window of the program
}
\usage{
MultBiplotRGUI(Toolkit = "RGtk2", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Toolkit}{
RGtk2. Not tried with other toolkits
}
  \item{\dots}{
Aditional parameters
}
}
\details{
Graphical user interface for the initial window of the program. Controls the data used in susequent calculations.
}
\value{
No value Returned.
%% ...
}

\author{
Jose Luis Vicente Villardon
}

\examples{
MultBiplotRGUI()
}

