\name{PCA.BiplotGUI}
\alias{PCA.BiplotGUI}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Graphical user interface for Classical Biplots
}
\description{
Graphical user interface for Classical Biplots
}
\usage{
PCA.BiplotGUI(X, Toolkit = "RGtk2", ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{
The Data Frame for the Biplot.
}
  \item{Toolkit}{
RGtk2. Not tried with other toolkits
}
  \item{\dots}{
Aditional Arguments
}
}
\details{
Graphical user interface for Classical Biplots
}
\value{
No value returned
}

\author{
Jose Luis Vicente Villardon
}


\examples{
PCA.BiplotGUI(iris)
}

